--- Taras_Martyniuk/4/log_config.json
@@ -1,58 +1,0 @@
-{ 
-    "version" : 1,
-    "disable_existing_loggers" : "False",
-    "formatters" : { 
-        "standard" : { 
-            "format" :  "[%(levelname)s]: %(message)s"
-        },
-        "precise" : {
-             "format" :  "%(asctime)s [%(levelname)s] %(name)s : %(message)s"
-        },
-        "lean" : {
-            "format" : "%(message)s"
-        }
-    },
-     "handlers" : { 
-         "console" : { 
-             "level" :  "DEBUG",
-             "formatter" :  "standard",
-             "class" :  "logging.StreamHandler",
-             "stream"  :  "ext://sys.stdout"
-        },
-         "file" : { 
-             "level" :  "DEBUG",
-             "formatter" :  "standard",
-             "class" :  "logging.handlers.RotatingFileHandler",
-             "filename"  :  "dungeon_logs.log"
-        },
-        "console_lean" : { 
-            "level" :  "INFO",
-            "formatter" :  "lean",
-            "class" :  "logging.StreamHandler",
-            "stream"  :  "ext://sys.stdout"
-       },
-       "file_lean" : { 
-        "level" :  "DEBUG",
-        "formatter" :  "lean",
-        "class" :  "logging.handlers.RotatingFileHandler",
-        "filename"  :  "dungeon_logs.log"
-        }
-    },
-     "loggers" : { 
-         "Debug Log File" : { 
-             "handlers" : ["file"],
-             "level" :  "DEBUG",
-             "propagate" : "True"
-        },
-        "Debug Log File Console" : { 
-            "handlers" : ["file", "console"],
-            "level" :  "DEBUG",
-            "propagate" : "True"
-       },
-        "Output Log" : {
-            "handlers" : [ "console_lean",  "file_lean"],
-            "level" :  "INFO",
-            "propagate" : "True"
-        }
-    } 
-}
