+++ Taras_Martyniuk/4/tm_dungeon_game/setup.py
@@ -1,0 +1,11 @@
+import setuptools
+
+with open("README.md", "r") as fh:
+    long_description = fh.read()
+
+setuptools.setup(
+    name="tm_dungeon_game",
+    version="0.0.1",
+    author="Taras Martyniuk",
+)
+
