+++ Taras_Martyniuk/4/tm_dungeon_game/tm_dungeon_game/__init__.py
@@ -1,0 +1,5 @@
+name = "tm_dungeon_game"
+
+
+
+
